+++ NEWS.md
@@ -1,5 +1,10 @@
 # see (devel)
 
+## Changes
+
+* Added plot method for prior predictive checks, from the `performance::check_priors()`
+  function.
+
 ## Bug fixes
 
 * The `type` argument in `performance::check_predictons()` was not recognized in